+++ packages/components/src/components/textarea/textarea-styles.spec.ts
@@ -3,20 +3,20 @@ import { getComponentCss } from './textarea-styles';
 
 describe('getComponentCss()', () => {
   it.each<Parameters<typeof getComponentCss>>([
-    [false, false, false, 'none', true, true, 'vertical', 'fixed', 'light'],
-    [false, true, false, 'none', true, true, 'vertical', 'fixed', 'light'],
-    [false, false, false, 'none', false, true, 'vertical', 'fixed', 'light'],
-    [false, false, true, 'none', true, true, 'vertical', 'fixed', 'light'],
-    [false, false, true, 'none', false, true, 'vertical', 'fixed', 'light'],
-    [false, false, false, 'success', true, true, 'vertical', 'fixed', 'light'],
-    [false, false, false, 'success', false, false, 'vertical', 'fixed', 'light'],
-    [false, false, true, 'success', true, false, 'vertical', 'content', 'light'],
-    [false, false, true, 'success', false, false, 'vertical', 'content', 'light'],
-    [false, false, false, 'error', true, false, 'vertical', 'fixed', 'light'],
-    [false, false, false, 'error', false, false, 'vertical', 'content', 'light'],
-    [false, false, true, 'error', true, false, 'vertical', 'fixed', 'light'],
-    [false, false, true, 'error', false, false, 'vertical', 'content', 'light'],
-    [true, false, true, 'error', false, false, 'vertical', 'fixed', 'light'],
+    [false, false, false, 'none', true, true, 'vertical', 'light'],
+    [false, true, false, 'none', true, true, 'vertical', 'light'],
+    [false, false, false, 'none', false, true, 'vertical', 'light'],
+    [false, false, true, 'none', true, true, 'vertical', 'light'],
+    [false, false, true, 'none', false, true, 'vertical', 'light'],
+    [false, false, false, 'success', true, true, 'vertical', 'light'],
+    [false, false, false, 'success', false, false, 'vertical', 'light'],
+    [false, false, true, 'success', true, false, 'vertical', 'light'],
+    [false, false, true, 'success', false, false, 'vertical', 'light'],
+    [false, false, false, 'error', true, false, 'vertical', 'light'],
+    [false, false, false, 'error', false, false, 'vertical', 'light'],
+    [false, false, true, 'error', true, false, 'vertical', 'light'],
+    [false, false, true, 'error', false, false, 'vertical', 'light'],
+    [true, false, true, 'error', false, false, 'vertical', 'light'],
     [
       false,
       false,
@@ -25,7 +25,6 @@ describe('getComponentCss()', () => {
       true,
       false,
       'vertical',
-      'content',
       'light',
     ],
     [
@@ -36,11 +35,10 @@ describe('getComponentCss()', () => {
       true,
       false,
       'vertical',
-      'fixed',
       'dark',
     ],
   ])(
-    'should return correct css for isDisabled: %s, isReadonly: %s, hideLabel: %o, state: %s, compact: %s, counter: %s, resize: %s, fieldSizing: %s, theme: %s',
+    'should return correct css for isDisabled: %s, isReadonly: %s, hideLabel: %o, state: %s, compact: %s, counter: %s, resize: %s, theme: %s',
     (...args) => {
       validateCssAndMatchSnapshot(getComponentCss(...args));
     }